Move Test suite to gem
