+++ gem/phlexing.gemspec
@@ -32,6 +32,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency "erb_parser", "~> 0.0.2"
   spec.add_dependency "html_press", "~> 0.8.2"
   spec.add_dependency "nokogiri", "~> 1.0"
+  spec.add_dependency "phlex", "~> 1.1"
   spec.add_dependency "rufo", "~> 0.13.0"
   spec.add_dependency "syntax_tree"
 end
