+++ gem/test/phlexing_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require "test_helper"
+require_relative "./test_helper"
 
-class TestPhlexing < Minitest::Test
-  def test_that_it_has_a_version_number
+class PhlexingTest < Minitest::Spec
+  it "has a version number" do
     refute_nil ::Phlexing::VERSION
   end
 end
